Gifting Agent & Blob Spawner
Implemented last of the three agents, the gifting agent. It can steal blobs from the house with the highest score, and will gift blobs to another house if it has a lower score. Changed all the state indicator colors to ones that fit better with the colors already in the game. Also added a script to spawn blobs wherever you click and implemented a delete feature when clicking on agents, which will then respawn them at their respective house.